Fixed compiler warnings: 103 unnecessary casts (CASTOR-1393)

+++ trunk/castor/src/main/org/exolab/castor/mapping/AccessMode.java
@@ -199,7 +199,7 @@ public class AccessMode implements Cloneable, Comparable, Serializable {
      * @return Hash code of this object.
      */
     public int hashCode() {
-        return (int) _id;
+        return _id;
     }
 
     /**
